+++ backend/packages/caching/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cat-factory/caching
 
+## 0.11.19
+
+### Patch Changes
+
+- Updated dependencies [4ecb25c]
+  - @cat-factory/kernel@0.190.0
+
+## 0.11.18
+
+### Patch Changes
+
+- Updated dependencies [7ed2bc0]
+  - @cat-factory/kernel@0.189.0
+
 ## 0.11.17
 
 ### Patch Changes