+++ backend/packages/gitlab/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @cat-factory/gitlab
 
+## 0.14.2
+
+### Patch Changes
+
+- Updated dependencies [4ecb25c]
+  - @cat-factory/kernel@0.190.0
+
+## 0.14.1
+
+### Patch Changes
+
+- Updated dependencies [7ed2bc0]
+  - @cat-factory/contracts@0.194.0
+  - @cat-factory/kernel@0.189.0
+
 ## 0.14.0
 
 ### Minor Changes